size_t s21_strlen(const char *str)
 - Release function size_t s21_strlen(const char *str)
 - Add unit test for function size_t s21_strlen(const char *str)
 - Fix data type size_t to "unsigned long" in x64 arc

+++ src/s21_string.c
@@ -101,8 +101,8 @@ TODO: Part 4. Дополнительно. Реализация функции ss
 модификаторов и типов преобразования).
 */
 
-int s21_strlen(const char *str) {
-  int len = 0;
+size_t s21_strlen(const char *str) {
+  size_t len = 0;
 
   for (; *(str + len); len++)
     ;
